+++ src/xrGame/alife_dynamic_object.cpp
@@ -311,7 +311,7 @@ void CSE_ALifeCar::add_offline(const xr_vector<ALife::_OBJECT_ID>& saved_childre
 	CSE_ALifeDynamicObjectVisual* object = (this);
 
 
-	for (u32 i = 0, n = saved_children.size(); i < n; ++i) {
+	for (u32 i = 0, n = (u32)saved_children.size(); i < n; ++i) {
 		CSE_ALifeDynamicObject* child = smart_cast<CSE_ALifeDynamicObject*>(ai().alife().objects().object(saved_children[i], true));
 		R_ASSERT2(child, make_string<const char*>("parent [%d][%s][%s] has not valid child [%d]",
 			ID, name(), name_replace(), saved_children[i]));
